[HER-170] Follow-up Rename the folder first-section
