+++ src/home/content.jsx
@@ -3,8 +3,8 @@ import { Grid } from '@material-ui/core'
 
 import { Marketing } from 'home/marketing'
 import { MeetTeam } from 'home/meet-team'
-import { useGridStyle } from 'home/first-section/hooks/use-grid-style'
-import { HeadingSection } from 'home/first-section'
+import { useGridStyle } from 'home/heading-section/hooks/use-grid-style'
+import { HeadingSection } from 'home/heading-section'
 
 export const Content = () => <Grid container>
   <HeadingSection />
