+++ src/home/heading-section/description/hooks/use-content-style.js
