+++ src/home/heading-section/hooks/use-style.js
@@ -1,7 +1,7 @@
 import { useMediaQuery, useTheme } from '@material-ui/core'
 
-import { useDesktopStyle } from 'home/first-section/hooks/use-desktop-style'
-import { usePhoneStyle } from 'home/first-section/hooks/use-phone-style'
+import { useDesktopStyle } from 'home/heading-section/hooks/use-desktop-style'
+import { usePhoneStyle } from 'home/heading-section/hooks/use-phone-style'
 
 export const useStyle = () => {
   const isMobile = useMediaQuery(useTheme().breakpoints.down('md'))
