+++ src/home/heading-section/link/content.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Grid, Link, Typography } from '@material-ui/core'
 
-import { useTitleStyle } from 'home/first-section/link/hooks/use-content-first-style'
-import LinkIcon from 'home/first-section/link/link_black_24dp.svg'
+import { useTitleStyle } from 'home/heading-section/link/hooks/use-content-first-style'
+import LinkIcon from 'home/heading-section/link/link_black_24dp.svg'
 import { useStyles } from 'core/layout/footer/content/style'
-import { useContentStyle } from 'home/first-section/link/hooks/use-content-style'
+import { useContentStyle } from 'home/heading-section/link/hooks/use-content-style'
 
 export const Content = () => <Grid item>
   <Typography classes={useTitleStyle()} color={'primary'}>git clone </Typography>
