+++ src/home/heading-section/link/hooks/use-content-first-style.js
@@ -1,6 +1,6 @@
 import { makeStyles } from '@material-ui/core'
 
-import { contentStyle } from 'home/first-section/link/content-style'
+import { contentStyle } from 'home/heading-section/link/content-style'
 
 export const useTitleStyle = makeStyles({
   root: {
