+++ src/home/heading-section/link/title.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Grid, Typography } from '@material-ui/core'
 
-import { useTitleStyle } from 'home/first-section/link/hooks/use-title-style'
+import { useTitleStyle } from 'home/heading-section/link/hooks/use-title-style'
 
 export const Title = () => <Grid item>
   <Typography color={'primary'} classes={useTitleStyle()}>
