+++ src/home/heading-section/live/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Grid } from '@material-ui/core'
 
-import { useStyle } from 'home/first-section/live/hooks/use-style'
-import { SeeButton } from 'home/first-section/live/see-button'
+import { useStyle } from 'home/heading-section/live/hooks/use-style'
+import { SeeButton } from 'home/heading-section/live/see-button'
 
 export const Live = () => <Grid item classes={useStyle()} justify={'left'}>
   <Grid container justify={'left'}>
