+++ src/home/heading-section/live/see-button/hooks/use-button-style.js
@@ -1,0 +1,9 @@
+import { useCamerasStyle } from 'home/heading-section/live/see-button/hooks/use-cameras-style'
+import { useDetectionsStyle } from 'home/heading-section/live/see-button/hooks/use-detections-style'
+
+export const useButtonStyle = path => {
+  const camerasStyle = useCamerasStyle()
+  const detectionsStyle = useDetectionsStyle()
+
+  return path === '/cameras' ? camerasStyle : detectionsStyle
+}
